OneCodeTeam/How to setup Vue.js with TypeScript in ASP.NET Core/[C#]-How to setup Vue.js with TypeScript in ASP.NET Core/C#/VueSource/src/components/navbar/navbar.ts
import Vue from 'vue';
import { Component, watch } from 'vue-property-decorator';
import { Link } from './link';
import { Logger } from '../../util/log';

@Component({
    template: require('./navbar.html')
})
export class NavbarComponent extends Vue {

    protected logger: Logger;

    inverted: boolean = true; // default value

    object: { default: string } = { default: 'Default object property!' }; // objects as default values don't need to be wrapped into functions

    links: Link[] = [
        new Link('Home', '/'),
        new Link('About', '/about'),
        new Link('List', '/list')
    ];

    @watch('$route.path')
    pathChanged() {
        this.logger.info('Changed current path to: ' + this.$route.path);
    }

    mounted() {
        if (!this.logger) this.logger = new Logger();
        this.$nextTick(() => this.logger.info(this.object.default));
    }
}
